rename types.rs to cli_types.rs and add more example
