+++ crates/rooch/src/commands/account/mod.rs
@@ -1,7 +1,7 @@
 // Copyright (c) RoochNetwork
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::types::CommandAction;
+use crate::cli_types::CommandAction;
 use async_trait::async_trait;
 use commands::{create::CreateCommand, import::ImportCommand, list::ListCommand};
 use rooch_types::error::{RoochError, RoochResult};
