+++ crates/rooch/src/commands/move_cli/commands/run_function.rs
@@ -1,7 +1,7 @@
 // Copyright (c) RoochNetwork
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::types::{ArgWithType, CommandAction, TransactionOptions, WalletContextOptions};
+use crate::cli_types::{ArgWithType, CommandAction, TransactionOptions, WalletContextOptions};
 use async_trait::async_trait;
 use clap::Parser;
 use moveos_types::{move_types::FunctionId, transaction::MoveAction};
@@ -32,7 +32,7 @@ pub struct RunFunction {
 
     /// Arguments combined with their type separated by spaces.
     ///
-    /// Supported types [u8, u16, u32, u64, u128, u256, bool, object_id, string, address, raw, vector<inner_type>]
+    /// Supported types [u8, u16, u32, u64, u128, u256, bool, object_id, string, address, vector<inner_type>]
     ///
     /// Example: `address:0x1 bool:true u8:0 u256:1234 'vector<u32>:a,b,c,d'`
     ///     address and uint can be written in short form like `@0x1 1u8 4123u256`.
