+++ crates/rooch/src/commands/transaction/mod.rs
@@ -1,7 +1,7 @@
 // Copyright (c) RoochNetwork
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::types::CommandAction;
+use crate::cli_types::CommandAction;
 use async_trait::async_trait;
 use commands::{get_by_hash::GetByHashCommand, get_by_index::GetByIndexCommand};
 use rooch_types::error::RoochResult;
